hue/requirements.txt
asgiref==3.3.1
coreapi==2.3.3
drf-spectacular==0.13.2
gunicorn==20.0.4
Django==3.1.6
django-cors-headers==3.7.0
django-celery-beat==1.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
mysqlclient==2.0.3
opentracing==2.4.0
pytz==2021.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
